shout2send: error out if no caps were received
Instead of assuming that input is ogg.

+++ ext/shout2/gstshout2.h
@@ -60,7 +60,7 @@ struct _GstShout2send {
   gchar *songmetadata;
   gchar *songartist;
   gchar *songtitle;
-  guint16 format;
+  int    format;
 
   GstTagList* tags;
 };
